requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
redis>=4.6.0
python-dotenv>=1.0.1
pytest>=8.3.0